agents/requirements.txt
# Solana Immune System - Python Agent Dependencies

# Anthropic Claude API
anthropic>=0.40.0

# Solana
solana>=0.34.0
anchorpy>=0.20.0
solders>=0.21.0

# Async
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Environment
python-dotenv>=1.0.0

# Logging
structlog>=24.0.0

# HTTP client
httpx>=0.27.0

# Crypto
pynacl>=1.5.0

# Scheduling
apscheduler>=3.10.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
